add named_group.rb && mv module TTTLS13::Message::Extension::NamedGroup => TTTLS13::NamedGroup
update server; add valid_settings?
