+++ lib/tttls1.3.rb
@@ -10,6 +10,7 @@ require 'tttls1.3/logging'
 require 'tttls1.3/error'
 require 'tttls1.3/cipher_suites'
 require 'tttls1.3/signature_scheme'
+require 'tttls1.3/named_group'
 require 'tttls1.3/cryptograph'
 require 'tttls1.3/transcript'
 require 'tttls1.3/key_schedule'
